Refactor services backend
